Version 1.3'

+++ html/upload.php
@@ -9,11 +9,16 @@ if($target_file=="html.zip"){
 	else{
 		echo "Could not upload file";
 	}
-	exec("unzip -P 'apexplus' -o /var/www/uploads/html.zip -d /var/www/uploads/");
+	exec("unzip -P 'apexplus' -o /var/www/uploads/html.zip -d /var/www/uploads/",$output,$result);
+	if($result==0){
 	exec("cp -TRv /var/www/uploads/html/ /var/www/");
 	exec("rm --interactive=never /var/www/uploads/html.zip");
 	exec("rm --interactive=never -r /var/www/uploads/html");
 	echo "Software has been installed";
+	}
+	else{
+		echo "Error unzipping the file";
+	}
 }
 else{
 	echo("Invalid file name");
